requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ECG Data Loading
wfdb>=4.1.0

# Configuration Management
pyyaml>=6.0

# Experiment Tracking & Visualization
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation Metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.7; python_version < "3.4"

